base64: move it to private for websockets
There are no other consumers for this, and reasonably unlikely to
be others for now.  (Other use cases are JWTs, but that would be
another whole set of functionality that we're not ready to take on.)
