+++ src/supplemental/websocket/base64.c
